# BUG/PERF: MaskedArray.__setitem__ validation