- [x] closes #44172
- [x] closes #44186
- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [ ] whatsnew entry

Fixes a bunch of ways we currently allow invalid setting, but ATM does _not_ fix the non-scalar equivalents.

Perf example from #44172 OP
```
df = pd.DataFrame([[True, False, pd.NA] * 200] * 200, dtype = "object")
df2 = df.astype("boolean")

%timeit res = df2.apply(lambda row: row.count(), axis = 1)
3.9 s ± 192 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- main
886 ms ± 40.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```