@jreback 2022-01-17
> > what's the bug fix here?
> 
> ```
> arr = pd.array([True, False])
> arr[0] = pd.NaT  # <- fails to raise
> ```

kk prob worth a note in a followon