@jreback 2022-01-17
@jbrockmendel seems something related to setitem is breaking on master: https://dev.azure.com/pandas-dev/pandas/_build/results?buildId=72425&view=logs&j=e2350215-d973-503e-a41e-d287ff823797&t=751e993b-1954-5330-832a-a54bd1064f7d&l=44